made userid an int, which ended up touching a lot
Former-commit-id: 60ca7913af5e236a6d015e3bd761629d75930d73

+++ src/mitll/langtest/client/custom/dialog/EditItem.java
@@ -277,7 +277,7 @@ public class EditItem {
    * @param userid
    * @return
    */
-  private UserExercise createNewItem(long userid) {
+  private UserExercise createNewItem(int userid) {
     return new UserExercise(-1, UserExercise.CUSTOM_PREFIX + Long.MAX_VALUE, userid, "", "", "");
   }
